Define a new method with(_:) on HTTPManagerRequest
This new method executes a block with `self` as the argument, and then
returns `self`. It's meant to aid in functional-style chaining.

Unfortunately, because Swift does not allow you to use `Self`
everywhere, every `HTTPManagerRequest` subclass has to override the
method just to provide a more specific argument type to the block.