Add methods to create tasks without automatically resuming them
This adds the methods `createTaskWithCompletion(_:)` and
`createTaskWithCompletionOnQueue(_:handler:)`. They're identical to the
old `performRequestWithCompletion(_:)` and
`performRequestWithCompletionOnQueue(_:handler:)` methods except they
don't automatically call `resume()` on the underlying
`NSURLSessionTask`.

This also adds the convenience methods `resume()` and `suspend()` to
`HTTPManagerTask`, which just call the equivalent method on the
underlying `NSURLSessionTask`.

The reason for this addition is for cases where the calling code needs
to record something about the `HTTPManagerTask` or `NSURLSessionTask`
prior to the completion block being invoked, and isn't already holding
the lock / serial queue used to protect access to the shared data.