Disable the disk cache for the tests
After implementing the mocking tests, the tests started intermittently
failing in a bizarre fashion, where the network tasks started timing out
without even trying to connect to the server. It turns out this is due
to an issue with the shared URL cache; the system console was showing
logs indicating that the disk cache database was unexpectedly locked.
Using an in-memory cache only works around this issue (because it won't
hit the disk at all).