Enforce no caching of POST/PUT/PATCH/DELETE requests
Make `cachePolicy` read-only at the `HTTPManagerRequest` level and
provide a readwrite version on `HTTPManagerDataRequest`. Set the
`cachePolicy` to `.ReloadIgnoringLocalCacheData` for
`HTTPManagerActionRequest`. `NSURLSession` should already be skipping
the cache for these requests, but we just want to be explicit about it.

This does mean that you cannot change the `cachePolicy` for parse
requests anymore, but you can always change the cache policy on the
data request before creating the parse request from it.