Rework how mocks compare their paths
The old way wasn't handling mocks with paths like `"/foo"` properly and
it wasn't handling %2f correctly (%2f is the percent-encoding for "/").
It also didn't handle paths like `"/:/foo"` correctly.

The new way compares path components, and it parses out the `:name`
tokens prior to converting the mock path into an absolute URL so that
way an environment with a path like `"/:foo"` wouldn't inadvertently
treat that as a token.