Replace the NSInputStream code with a new implementation
We can't use the private `CFReadStream` APIs from CFStreamAbstract.h
because of app review. So the whole thing has been reimplemented as a
custom `NSInputStream` subclass. All of the relevant documentation on
the internet indicates that this approach requires overriding private
APIs and is less reliable than the `CFReadStream` approach, but in my
testing this is no longer true (at least as of iOS 8.0). Our stream
works correctly without any private API, although it did turn out to be
a huge amount of work to implement the stream properly and safely.