Ditch the private module and just hide the classes
It turns out that the way we were doing a private module before caused
the full path to the module to be embedded in the framework binary. This
doesn't affect its usage at runtime, but it is used by any other code
that imports the framework. The result of this was that removing the
Carthage checkout from the dependent project, or copying the built
frameworks to another computer (e.g. with carthage_cache), caused the
build of the dependent project to fail.

Unfortunately there doesn't seem to be any way to define the private
module such that it solves the above problem without the symbols also
implicitly being visible to dependent projects who import the framework.
Given this limitation, we've now just renamed the private types to
start with `_` and added `__attribute__((visibility("hidden")))` to the
private classes. Dependents can still reference the private types, but
any reference to the classes will fail to link.

We're also using a manual module.modulemap file for good measure, since
that lets us split the headers cleanly, and maybe someday the compiler
will actually start respecting the explicit submodule.