Change how the network activity tracking works
PMHTTP no longer sets `UIApplication. networkActivityIndicatorVisible`
itself. Instead it vends a static property which is a block that is
invoked (on the main thread) every time the number of outstanding tasks
changes. This way the calling app can then use that to determine whether
to show the network activity indicator. This change allows PMHTTP to
coexist with other networking libraries, and it also removes the main
reason why PMHTTP wasn't marked as extension-safe.

Also fix the issue where we were showing the network activity indicator
as soon as a task was created, even if the task hadn't been started yet.
This means we need to be more careful about tracking the state.

Remove the method `HTTPManager.trackNetworkActivityForTask(_:)`. Now
that we're not mucking with the network activity indicator directly,
it's the application's job to deal with tracking any external tasks.