Skip the `HTTPManagerConfigurable` support in application extensions
The call to `UIApplication.sharedApplication()` was the last remaining
non-extension-safe API. There's no simple way to test if we're in an
extension, so we use the slightly hacky mechanism of checking if the
main bundle's Info.plist has the `"NSExtension"` key. If the key is not
present, we assume we're in an application. We still can't call
`UIApplication.sharedApplication()` directly, but we can access it with
`valueForKey(_:)`.

With this change, we can now set `APPLICATION_EXTENSION_API_ONLY`.