Ensure all calls to expectation.fulfill() happen on the main thread
XCTest doesn't document anything to do with thread safety, so just to be
sure, let's always call fulfill() from the main thread. Also modify all
these calls to hold the expectation weakly. This way if the code runs
after the timeout has occurred, the weak expectation will be nil and we
won't crash the test runner.