Encode application/x-www-form-urlencoded better
`application/x-www-form-urlencoded` defines `+` as meaning a space. All
servers that accept this content type should interpret it that way, and
most servers also interpret query strings the exact same way.
Unfortunately, we weren't encoding `+` at all. And it turns out that
`URLComponents.queryItems` also doesn't treat `+` specially.

In the interests of maximum compatibility, we actually convert spaces to
`%20` instead of to `+`.