Cancel zombie tasks when the session is invalidated
If the user creates a task but never resumes it, when the session is
invalidated the task ends up as a zombie. When this happens, we should
call the completion block to inform it that it was canceled. The
underlying network task won't have been canceled, but by the time the
session is invalidated, canceling that does nothing.