Change how we serialize nested array/set parameters
Don't automatically add the `"[]"` onto the key, because servers may in
fact just want to have the key repeated several times. The user can put
the `"[]"` on the key themselves if necessary. If such a key is found in
a dictionary or nested `URLQueryItem`, move the trailing `"[]"` outside
of the `"dict[key]"` brackets, so we end up with e.g. `"dict[key][]"`.