Add new property serverRequiresContentLength
This property turns off body streaming and causes us to encode JSON and
multipart bodies synchronously instead, so we can provide a
Content-Length.

As usual, there is a corresponding `defaultServerRequiresContentLength`
property on the `HTTPManager`, which only applies to requests contained
within the environment.

Add a method `HTTPManagerRequest.setDefaultEnvironmentalProperties()`.
This sets `auth` and `serverRequiresContentLength` to the `HTTPManager`
defaults regardless of the environment.